Renamed `firebaseAdmin` + other type cleanup
Also added `future` flags to React Router to silence console logs
