+++ package.json
@@ -25,14 +25,14 @@
     "ci": "yarn install --force",
     "build": "yarn workspaces run build",
     "build:shared": "yarn workspace @conductor/shared build",
-    "build:pwa": "yarn run build:shared && yarn workspace @conductor/pwa build",
-    "build:scripts": "yarn run build:shared && yarn workspace @conductor/scripts build",
-    "build:server": "yarn run build:shared && yarn workspace @conductor/webServer build",
-    "build:extension": "yarn run build:shared && yarn workspace @conductor/extension build",
-    "build:functions": "yarn run build:shared && yarn workspace @conductor/functions build",
+    "build:pwa": "yarn workspace @conductor/shared build && yarn workspace @conductor/pwa build",
+    "build:scripts": "yarn workspace @conductor/shared build && yarn workspace @conductor/scripts build",
+    "build:server": "yarn workspace @conductor/shared build && yarn workspace @conductor/webServer build",
+    "build:extension": "yarn workspace @conductor/shared build && yarn workspace @conductor/extension build",
+    "build:functions": "yarn workspace @conductor/shared build && yarn workspace @conductor/functions build",
     "firebase:local": "firebase emulators:start",
-    "deploy:pwa": "yarn run build:shared && yarn workspace @conductor/pwa deploy",
-    "deploy:functions": "yarn run build:shared && yarn workspace @conductor/functions deploy",
+    "deploy:pwa": "yarn workspace @conductor/shared build && yarn workspace @conductor/pwa deploy",
+    "deploy:functions": "yarn workspace @conductor/shared build && yarn workspace @conductor/functions deploy",
     "deploy:rules": "firebase deploy --only firestore:rules,storage",
     "deploy:rules:firestore": "firebase deploy --only firestore:rules",
     "deploy:rules:storage": "firebase deploy --only storage",
