+++ packages/functions/src/lib/feedSubscriptions.ts
@@ -3,7 +3,7 @@ import {FEED_SUBSCRIPTIONS_DB_COLLECTION} from '@shared/lib/constants';
 import {FeedSubscription} from '@shared/types/feedSubscriptions.types';
 import {UserId} from '@shared/types/user.types';
 
-import {firestore} from '@src/lib/firebase';
+import {firestore} from '@src/lib/firebaseAdmin';
 
 import {batchDeleteFirestoreDocuments} from './batch';
