+++ packages/functions/src/lib/users.ts
@@ -2,7 +2,7 @@ import {USERS_DB_COLLECTION} from '@shared/lib/constants';
 
 import {UserId} from '@shared/types/user.types';
 
-import {firestore} from '@src/lib/firebase';
+import {firestore} from '@src/lib/firebaseAdmin';
 
 /**
  * Hard-deletes a user document from Firestore.
